fix: fix issue for recon service
